Spawn food on map, spawn bots when food is taken
To take a piece of food, just move onto the tile where the food resides.
Food pieces will randomly spawn at walkable tiles from time to time.

When taking a food piece with a bot, a new bot will spawn for that
commander on the home base, provided the home base is not occupied. If
a bot can be spawned, the food piece will be consumed, else it remains
on the ground.